requirements-docker.txt
pandas
sqlalchemy
sqlalchemy-iris